Move option helpers into utils
