+++ libs/llrt_utils/src/lib.rs
@@ -14,6 +14,7 @@ pub mod macros;
 pub mod mc_oneshot;
 pub mod module;
 pub mod object;
+pub mod option;
 pub mod primordials;
 pub mod result;
 pub mod reuse_list;
