+++ modules/llrt_stream_web/src/lib.rs
@@ -1,5 +1,6 @@
 use llrt_utils::{
     module::{export_default, ModuleInfo},
+    option::Undefined,
     primordials::Primordial,
 };
 use queuing_strategy::{ByteLengthQueuingStrategy, CountQueuingStrategy, SizeValue};
@@ -16,7 +17,7 @@ use rquickjs::{
     prelude::{IntoArg, OnceFn, This},
     promise::PromiseState,
     Class, Ctx, Error, Exception, FromJs, Function, IntoAtom, IntoJs, JsLifetime, Object, Promise,
-    Result, Type, Value,
+    Result, Value,
 };
 use std::collections::VecDeque;
 use std::{cell::Cell, rc::Rc};
@@ -120,68 +121,6 @@ pub fn init(ctx: &Ctx) -> Result<()> {
     Class::<WritableStreamDefaultController>::define(globals)?;
 
     Ok(())
-}
-
-/// Helper type for treating an undefined value as None, but not null
-#[derive(Clone)]
-struct Undefined<T>(pub Option<T>);
-
-impl<'js, T: FromJs<'js>> FromJs<'js> for Undefined<T> {
-    fn from_js(ctx: &Ctx<'js>, value: Value<'js>) -> Result<Self> {
-        if value.type_of() == Type::Undefined {
-            Ok(Self(None))
-        } else {
-            Ok(Self(Some(FromJs::from_js(ctx, value)?)))
-        }
-    }
-}
-
-impl<T> Default for Undefined<T> {
-    fn default() -> Self {
-        Self(None)
-    }
-}
-
-unsafe impl<'js, T: JsLifetime<'js>> JsLifetime<'js> for Undefined<T> {
-    type Changed<'to> = Undefined<T::Changed<'to>>;
-}
-
-impl<'js, T: Trace<'js>> Trace<'js> for Undefined<T> {
-    fn trace<'a>(&self, tracer: Tracer<'a, 'js>) {
-        self.0.trace(tracer)
-    }
-}
-
-/// Helper type for converting an option into null instead of undefined.
-#[derive(Clone)]
-struct Null<T>(pub Option<T>);
-
-impl<'js, T: IntoJs<'js>> IntoJs<'js> for Null<T> {
-    fn into_js(self, ctx: &Ctx<'js>) -> Result<Value<'js>> {
-        match self.0 {
-            None => Ok(Value::new_null(ctx.clone())),
-            Some(val) => val.into_js(ctx),
-        }
-    }
-}
-
-unsafe impl<'js, T: JsLifetime<'js>> JsLifetime<'js> for Null<T> {
-    type Changed<'to> = Null<T::Changed<'to>>;
-}
-
-impl<'js, T: Trace<'js>> Trace<'js> for Null<T> {
-    fn trace<'a>(&self, tracer: Tracer<'a, 'js>) {
-        self.0.trace(tracer)
-    }
-}
-
-impl<'js, T: IntoJs<'js>> IntoJs<'js> for Undefined<T> {
-    fn into_js(self, ctx: &Ctx<'js>) -> Result<Value<'js>> {
-        match self.0 {
-            None => Ok(Value::new_undefined(ctx.clone())),
-            Some(val) => val.into_js(ctx),
-        }
-    }
 }
 
 fn class_from_owned_borrow_mut<'js, T: JsClass<'js>>(
